requirements.txt
GitPython==0.3.5
coloredlogs==0.8
tornado==4.1
checkio_docker==0.0.15
PyYAML
future
